+++ mutable_buffer/src/lib.rs
@@ -59,8 +59,6 @@
     clippy::clone_on_ref_ptr
 )]
 
-pub mod checkpoint;
 pub mod chunk;
 mod column;
 mod dictionary;
-pub mod persistence_windows;
